+++ src/main/java/org/folio/uk/integration/keycloak/RealmConfigurationProvider.java
@@ -1,9 +1,9 @@
 package org.folio.uk.integration.keycloak;
 
 import lombok.RequiredArgsConstructor;
-import org.folio.common.configuration.properties.FolioEnvironment;
 import org.folio.tools.store.SecureStore;
 import org.folio.tools.store.exception.NotFoundException;
+import org.folio.tools.store.properties.SecureStoreProperties;
 import org.folio.uk.integration.keycloak.config.KeycloakProperties;
 import org.folio.uk.integration.keycloak.model.KeycloakRealmConfiguration;
 import org.springframework.cache.annotation.CacheEvict;
@@ -15,9 +15,9 @@ import org.springframework.stereotype.Service;
 public class RealmConfigurationProvider {
 
   private static final String MASTER_REALM = "master";
-  private final FolioEnvironment folioEnvironment;
   private final SecureStore secureStore;
   private final KeycloakProperties keycloakConfigurationProperties;
+  private final SecureStoreProperties secureStoreProperties;
 
   /**
    * Provides realm configuration using {@link org.folio.spring.FolioExecutionContext} object.
@@ -50,7 +50,7 @@ public class RealmConfigurationProvider {
 
   private String retrieveKcClientSecret(String realm, String clientId) {
     try {
-      return secureStore.get(buildKey(folioEnvironment.getEnvironment(), realm, clientId));
+      return secureStore.get(buildKey(secureStoreProperties.getEnvironment(), realm, clientId));
     } catch (NotFoundException e) {
       throw new IllegalStateException(String.format(
         "Failed to get value from secure store [clientId: %s]", clientId), e);